Create functionality in Cause class
Created instance and class methods to help Cause class function properly. Initializes with an 'id' and 'name' and saves to the all class array. Wrote the #all getter method and #save instance method. Checked with fake data to make sure incoming hash is itererated through, creates instances of a cause for each with 'id' and 'name', and saves each to the all class array.