API setup for getting charities within cause
Created framework for #ask_for_charities and #get_charities_from_api methods in CLI class.
Testing to see what the best way to handle importing large amount of charities from API is. Looking at 1) displaying the names of the charities and only creating an instance of
the organization if the user selects it (could lead to a lot of API hits). 2) Creating instances of each charity at the first API hit and then displaying the name of each
instance (this could take a long time to create over 10k instances in some cases).