Restructured methods in API and CLI class
Workiing on flow simplifying flow and structure of ClI and API classes
to allow the flow of the application to be more logical. Also created an
api_key.rb file to hold other programmers api_key so they don't have to
worry about plugging it in the right place. Should just insert in the file
and be ready to use the API.